examples/content-recommender/scraper/requirements.txt
scrapy==2.11.0
kafka-python-ng==2.2.2
schedule==1.2.1
